requirements.txt
locust
fastapi
uvicorn
ultralytics
pillow
python-multipart

